Migrate expression operator diagnostics
