+++ docs/errors/SIFR-TYPE-0012.md
@@ -1,0 +1,16 @@
+# SIFR-TYPE-0012
+
+<!-- Generated by cargo run -p sifr_diagnostics --bin gen-error-docs. Do not edit by hand. -->
+
+Unsupported expression form.
+
+| Field | Value |
+| --- | --- |
+| Code | `SIFR-TYPE-0012` |
+| Family | `TYPE` |
+| Severity | Error |
+| Owner | `sifr_hir::lower::expressions` |
+| Message template | `unsupported expression form: {form}` |
+| Representative fixture | `crates/sifr/tests/e2e/fail/unsupported_yield_expression.sifr` |
+| Declared args | `form (message+json)` |
+| Dedupe args | `form` |
