+++ internal_docs/diagnostic_codes.md
@@ -84,6 +84,7 @@ Tooling metadata defaults: `tool_actions` is empty, `fix_all_eligible` is `false
 | `SIFR-TYPE-0009` | `TYPE` | Active | Error | `docs/errors/SIFR-TYPE-0009.md` | `crates/sifr/tests/e2e/fail/tuple_unpack_shape_mismatch.sifr` | `sifr_hir::lower::tuple_unpack` | `tuple unpacking: expected {expected_count} values, got {actual_count}` | `expected_count (message+json)`, `actual_count (message+json)` | `expected_count`, `actual_count` | n/a | false |
 | `SIFR-TYPE-0010` | `TYPE` | Active | Error | `docs/errors/SIFR-TYPE-0010.md` | `crates/sifr/tests/e2e/fail/typevar_constraints_violation.sifr` | `sifr_hir::lower::expressions` | `type '{actual}' does not satisfy constraints ({constraints}) required by type parameter '{type_param}'` | `actual (message+json)`, `constraints (message+json)`, `type_param (message+json)` | `actual`, `constraints`, `type_param` | n/a | false |
 | `SIFR-TYPE-0011` | `TYPE` | Active | Error | `docs/errors/SIFR-TYPE-0011.md` | `crates/sifr/tests/e2e/fail/unsupported_default_expr_call.sifr` | `sifr_hir::lower::typing_and_functions` | `function {function}: unsupported default argument expression for parameter {parameter}` | `function (message+json)`, `parameter (message+json)` | `function`, `parameter` | n/a | false |
+| `SIFR-TYPE-0012` | `TYPE` | Active | Error | `docs/errors/SIFR-TYPE-0012.md` | `crates/sifr/tests/e2e/fail/unsupported_yield_expression.sifr` | `sifr_hir::lower::expressions` | `unsupported expression form: {form}` | `form (message+json)` | `form` | n/a | false |
 | `SIFR-TYPE-0901` | `TYPE` | Active | Warning | `docs/errors/SIFR-TYPE-0901.md` | `crates/sifr_driver/src/tests/single_file_frontend.rs::test_type_check_source_surfaces_arithmetic_overflow_as_structured_warning` | `sifr_hir::lower::arithmetic_warnings` | `integer {operation} may overflow at runtime` | `operation (message+json)` | `operation` | n/a | false |
 | `SIFR-TYPE-0902` | `TYPE` | Active | Note | `docs/errors/SIFR-TYPE-0902.md` | `crates/sifr_driver/src/tests/single_file_frontend.rs::test_type_check_source_surfaces_reveal_type_as_structured_note` | `sifr_hir::lower::builtin_calls` | `revealed type is {revealed_type}` | `revealed_type (message+json)` | `revealed_type` | n/a | false |
 | `SIFR-DECIMAL-0001` | `DECIMAL` | Active | Error | `docs/errors/SIFR-DECIMAL-0001.md` | `crates/sifr/tests/e2e/fail/decimal_invalid_literal_string.sifr` | `sifr_hir::lower` | `invalid Decimal literal: {literal}` | `literal (message+json)` | `literal` | n/a | false |
